fix: address PR review feedback (P1/P2/P3 issues)
- Add disabled/loading state to PayoutConfigurationModal buttons
- Fix sidebar blanking on page change with ref-based fallback
- Extract DetailsSection and MilestonesSection into separate files
- Wrap both extracted components in React.memo
- Rename test file to match component rename (modal → sidebar)
- Remove unused hasAllocationsError variable
