+++ __tests__/control-center/unit/project-details-sidebar.test.tsx
