+++ components/Pages/Admin/ControlCenter/ProjectDetailsSidebar.tsx
@@ -3,18 +3,13 @@
 import {
   ArrowTopRightOnSquareIcon,
   CheckIcon,
-  ClipboardDocumentIcon,
   ClockIcon,
   Cog6ToothIcon,
-  ExclamationTriangleIcon,
   InformationCircleIcon,
   PlusCircleIcon,
-  XMarkIcon,
 } from "@heroicons/react/24/outline";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import toast from "react-hot-toast";
-import { KycStatusBadge } from "@/components/KycStatusIcon";
-import { DatePicker } from "@/components/Utilities/DatePicker";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
@@ -24,33 +19,28 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
 import { useCopyToClipboard } from "@/hooks/useCopyToClipboard";
 import {
   type CommunityPayoutAgreementInfo,
   type CommunityPayoutInvoiceInfo,
-  formatDisplayAmount,
   fromSmallestUnit,
   type MilestoneAllocation,
-  MilestoneLifecycleStatus,
-  type MilestonePaymentStatus,
   PayoutConfigurationContent,
   type PayoutConfigurationContentRef,
   type PayoutDisbursement,
   PayoutDisbursementStatus,
   PayoutHistoryContent,
-  TokenBreakdown,
   type TokenTotal,
   useSaveMilestoneInvoices,
   useToggleAgreement,
 } from "@/src/features/payout-disbursement";
 import type { KycStatusResponse } from "@/types/kyc";
-import { formatAddressForDisplay } from "@/utilities/donations/helpers";
-import { formatDate } from "@/utilities/formatDate";
 import { getChainNameById } from "@/utilities/network";
 import { PAGES } from "@/utilities/pages";
 import { cn } from "@/utilities/tailwind";
+import { DetailsSection } from "./DetailsSection";
+import { MilestonesSection } from "./MilestonesSection";
 
 // ─── Props ──────────────────────────────────────────────────────────────────
 
@@ -97,113 +87,6 @@ const NAV_ITEMS: { id: SidebarSection; label: string; icon: typeof InformationCi
   { id: "settings", label: "Payout Settings", icon: Cog6ToothIcon },
   { id: "history", label: "History", icon: ClockIcon },
 ];
-
-// ─── Payment status display config ──────────────────────────────────────────
-
-const paymentStatusConfig: Record<
-  MilestonePaymentStatus,
-  { label: string; dotColor: string; textColor: string }
-> = {
-  unpaid: {
-    label: "Unpaid",
-    dotColor: "bg-gray-300 dark:bg-zinc-600",
-    textColor: "text-gray-500 dark:text-zinc-500",
-  },
-  pending: {
-    label: "Pending",
-    dotColor: "bg-amber-400",
-    textColor: "text-amber-600 dark:text-amber-400",
-  },
-  awaiting_signatures: {
-    label: "Awaiting sigs",
-    dotColor: "bg-blue-400",
-    textColor: "text-blue-600 dark:text-blue-400",
-  },
-  disbursed: {
-    label: "Disbursed",
-    dotColor: "bg-green-500",
-    textColor: "text-green-600 dark:text-green-400",
-  },
-};
-
-const milestoneStatusConfig: Record<
-  MilestoneLifecycleStatus,
-  { label: string; dotColor: string; textColor: string }
-> = {
-  [MilestoneLifecycleStatus.PENDING]: {
-    label: "Pending",
-    dotColor: "bg-gray-300 dark:bg-zinc-600",
-    textColor: "text-gray-500 dark:text-zinc-500",
-  },
-  [MilestoneLifecycleStatus.COMPLETED]: {
-    label: "Completed",
-    dotColor: "bg-blue-400",
-    textColor: "text-blue-600 dark:text-blue-400",
-  },
-  [MilestoneLifecycleStatus.VERIFIED]: {
-    label: "Verified",
-    dotColor: "bg-green-500",
-    textColor: "text-green-600 dark:text-green-400",
-  },
-  [MilestoneLifecycleStatus.PAST_DUE]: {
-    label: "Past due",
-    dotColor: "bg-amber-500",
-    textColor: "text-amber-600 dark:text-amber-400",
-  },
-};
-
-const invoiceStatusConfig: Record<string, { label: string; className: string }> = {
-  not_submitted: {
-    label: "Not submitted",
-    className: "bg-gray-100 text-gray-600 dark:bg-zinc-800 dark:text-zinc-400",
-  },
-  received: {
-    label: "Invoice received",
-    className: "bg-emerald-50 text-emerald-700 dark:bg-emerald-900/30 dark:text-emerald-400",
-  },
-};
-
-function getEffectiveMilestoneStatus(
-  milestoneStatus: MilestoneLifecycleStatus | null,
-  milestoneDueDate: string | null
-): MilestoneLifecycleStatus {
-  const status = milestoneStatus || MilestoneLifecycleStatus.PENDING;
-  if (status === MilestoneLifecycleStatus.PENDING && milestoneDueDate) {
-    const due = new Date(milestoneDueDate);
-    if (due.getTime() < Date.now()) {
-      return MilestoneLifecycleStatus.PAST_DUE;
-    }
-  }
-  return status;
-}
-
-function getInvoiceStatusKey(invoiceStatus: string): string {
-  return invoiceStatus === "received" || invoiceStatus === "paid" ? "received" : "not_submitted";
-}
-
-function getMilestoneStatusTooltip(
-  effectiveStatus: MilestoneLifecycleStatus,
-  statusUpdatedAt: string | null,
-  dueDate: string | null
-): string {
-  const fmtDate = (iso: string) => formatDate(iso, "UTC");
-  switch (effectiveStatus) {
-    case MilestoneLifecycleStatus.COMPLETED:
-      return statusUpdatedAt ? `Completed on ${fmtDate(statusUpdatedAt)}` : "Completed";
-    case MilestoneLifecycleStatus.VERIFIED:
-      return statusUpdatedAt ? `Verified on ${fmtDate(statusUpdatedAt)}` : "Verified";
-    case MilestoneLifecycleStatus.PAST_DUE:
-      return dueDate ? `Due ${fmtDate(dueDate)}` : "Past due";
-    case MilestoneLifecycleStatus.PENDING: {
-      const parts: string[] = [];
-      if (statusUpdatedAt) parts.push(`Created ${fmtDate(statusUpdatedAt)}`);
-      if (dueDate) parts.push(`Due ${fmtDate(dueDate)}`);
-      return parts.length > 0 ? parts.join(" · ") : "Pending";
-    }
-    default:
-      return effectiveStatus;
-  }
-}
 
 // ─── Component ───────────────────────────────────────────────────────────────
 
@@ -714,509 +597,5 @@ export function ProjectDetailsSidebar({
         </DialogContent>
       </Dialog>
     </Dialog>
-  );
-}
-
-// ─── Details Section ──────────────────────────────────────────────────────────
-
-interface DetailsSectionProps {
-  grant: ProjectDetailsSidebarGrant;
-  kycStatus: KycStatusResponse | null;
-  agreement: CommunityPayoutAgreementInfo | null;
-  localAgreementSigned: boolean;
-  agreementDate: Date | undefined;
-  confirmingUnsign: boolean;
-  setConfirmingUnsign: (v: boolean) => void;
-  toggleAgreementMutation: ReturnType<typeof useToggleAgreement>;
-  handleSignAgreement: (date?: Date) => void;
-  handleUnsignAgreement: () => void;
-  setAgreementDate: (date: Date | undefined) => void;
-  handleCopyAddress: () => void;
-  totalsByToken: TokenTotal[];
-  remainingBalance: {
-    approved: number;
-    totalDisbursed: number;
-    remaining: number;
-    pct: number;
-  } | null;
-  awaitingTx: PayoutDisbursement | null;
-  chainInfo: { chainID: number; token: string; tokenDecimals: number } | null;
-  milestoneSummary: { total: number; received: number; paid: number } | null;
-  invoiceRequired: boolean;
-}
-
-function DetailsSection({
-  grant,
-  kycStatus,
-  agreement,
-  localAgreementSigned,
-  agreementDate,
-  confirmingUnsign,
-  setConfirmingUnsign,
-  toggleAgreementMutation,
-  handleSignAgreement,
-  handleUnsignAgreement,
-  setAgreementDate,
-  handleCopyAddress,
-  totalsByToken,
-  remainingBalance,
-  awaitingTx,
-  chainInfo,
-  milestoneSummary,
-  invoiceRequired,
-}: DetailsSectionProps) {
-  return (
-    <div className="space-y-4">
-      {/* Status panel */}
-      <div className="rounded-lg bg-gray-50 dark:bg-zinc-900 p-3 space-y-2.5">
-        {/* KYC + Agreement row */}
-        <div className="flex items-center gap-3 flex-wrap">
-          <div className="flex items-center gap-1.5">
-            <span className="text-xs font-medium text-gray-500 dark:text-zinc-400">KYC/KYB:</span>
-            <KycStatusBadge status={kycStatus} showValidityInLabel={false} />
-          </div>
-          <div className="flex items-center gap-2 px-2 py-1.5 rounded-md bg-white dark:bg-zinc-800/50">
-            <span className="text-xs font-medium text-gray-500 dark:text-zinc-400">Agreement:</span>
-            {confirmingUnsign ? (
-              <div className="flex items-center gap-2 px-2 py-0.5 rounded bg-amber-50 dark:bg-amber-900/20 border border-amber-200 dark:border-amber-800/50">
-                <span className="text-xs text-amber-700 dark:text-amber-300">
-                  Mark as unsigned?
-                </span>
-                <Button
-                  size="sm"
-                  variant="destructive"
-                  className="h-5 px-2 text-[11px]"
-                  onClick={handleUnsignAgreement}
-                  aria-label="Confirm unsign agreement"
-                >
-                  Confirm
-                </Button>
-                <Button
-                  size="sm"
-                  variant="ghost"
-                  className="h-5 px-2 text-[11px]"
-                  onClick={() => setConfirmingUnsign(false)}
-                  aria-label="Cancel unsign agreement"
-                >
-                  Cancel
-                </Button>
-              </div>
-            ) : localAgreementSigned ? (
-              <div className="flex items-center gap-1.5">
-                <span className="h-1.5 w-1.5 rounded-full bg-emerald-500 shrink-0" />
-                <span className="text-xs font-medium text-emerald-700 dark:text-emerald-400">
-                  Signed {agreementDate ? formatDate(agreementDate, "UTC") : ""}
-                </span>
-                {agreement?.signedBy && (
-                  <span
-                    className="text-[10px] text-gray-400 dark:text-zinc-500"
-                    title={agreement.signedBy}
-                  >
-                    by {formatAddressForDisplay(agreement.signedBy)}
-                  </span>
-                )}
-                <button
-                  type="button"
-                  onClick={() => setConfirmingUnsign(true)}
-                  disabled={toggleAgreementMutation.isPending}
-                  className="ml-0.5 p-0.5 rounded hover:bg-gray-100 dark:hover:bg-zinc-700 text-gray-400 dark:text-zinc-500 hover:text-gray-600 dark:hover:text-zinc-300 transition-colors disabled:opacity-50"
-                  aria-label="Remove agreement signed date"
-                  title="Mark as unsigned"
-                >
-                  <XMarkIcon className="h-3.5 w-3.5" />
-                </button>
-              </div>
-            ) : (
-              <div className="flex items-center gap-1.5">
-                <span className="text-xs text-gray-400 dark:text-zinc-500">Not signed</span>
-                <span className="text-[10px] text-gray-300 dark:text-zinc-600">&mdash;</span>
-                <DatePicker
-                  selected={agreementDate}
-                  onSelect={(date) => {
-                    setAgreementDate(date);
-                    handleSignAgreement(date);
-                  }}
-                  maxDate={new Date()}
-                  placeholder="Pick a date"
-                  buttonClassName="h-6 text-xs px-2 py-0.5 bg-white dark:bg-zinc-900"
-                  ariaLabel="Set agreement signed date"
-                />
-                {toggleAgreementMutation.isPending && (
-                  <span className="text-[11px] text-gray-400 dark:text-zinc-500 animate-pulse">
-                    Saving...
-                  </span>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Payout summary */}
-        <div className="flex items-center gap-5 flex-wrap text-xs text-gray-500 dark:text-zinc-400">
-          <div className="flex items-center gap-1.5">
-            <span className="font-medium">Address:</span>
-            {grant.currentPayoutAddress ? (
-              <button
-                type="button"
-                onClick={handleCopyAddress}
-                className="inline-flex items-center gap-1 font-mono hover:text-blue-600 dark:hover:text-blue-400 transition-colors cursor-pointer"
-                title={`Click to copy: ${grant.currentPayoutAddress}`}
-              >
-                {formatAddressForDisplay(grant.currentPayoutAddress)}
-                <ClipboardDocumentIcon className="h-3 w-3 opacity-50" />
-              </button>
-            ) : (
-              <span className="text-amber-600 dark:text-amber-400 italic font-sans">
-                Not configured
-              </span>
-            )}
-          </div>
-          <div className="flex items-center gap-1.5">
-            <span className="font-medium">Approved:</span>
-            <span className="tabular-nums text-sm text-gray-700 dark:text-zinc-300">
-              {grant.currentAmount && parseFloat(grant.currentAmount) > 0
-                ? `${formatDisplayAmount(grant.currentAmount)}${grant.currency ? ` ${grant.currency}` : ""}`
-                : "\u2014"}
-            </span>
-          </div>
-          {totalsByToken.length > 0 && (
-            <div className="flex items-center gap-1.5">
-              <span className="font-medium">Disbursed:</span>
-              <span className="text-sm">
-                <TokenBreakdown totalsByToken={totalsByToken} size="sm" />
-              </span>
-            </div>
-          )}
-        </div>
-
-        {/* Remaining balance progress */}
-        {remainingBalance && (
-          <div className="flex items-center gap-3">
-            <div className="flex-1 h-1.5 rounded-full bg-gray-200 dark:bg-zinc-700 overflow-hidden">
-              <div
-                className={cn(
-                  "h-full rounded-full transition-all",
-                  remainingBalance.pct >= 100
-                    ? "bg-green-500"
-                    : remainingBalance.pct >= 75
-                      ? "bg-blue-500"
-                      : "bg-blue-400"
-                )}
-                style={{ width: `${Math.min(100, remainingBalance.pct)}%` }}
-              />
-            </div>
-            <span className="text-[10px] tabular-nums text-gray-400 dark:text-zinc-500 whitespace-nowrap">
-              {remainingBalance.pct}% disbursed
-              {remainingBalance.remaining > 0 && (
-                <> · {formatDisplayAmount(String(remainingBalance.remaining))} remaining</>
-              )}
-            </span>
-          </div>
-        )}
-      </div>
-
-      {/* Awaiting signatures banner */}
-      {awaitingTx && (
-        <div className="flex items-center gap-2 px-3 py-2 rounded-lg bg-amber-50 dark:bg-amber-900/20 border border-amber-200 dark:border-amber-800/50">
-          <ExclamationTriangleIcon className="h-4 w-4 text-amber-600 dark:text-amber-400 shrink-0" />
-          <span className="text-xs text-amber-700 dark:text-amber-300">
-            Awaiting Safe signatures{" — "}
-            {formatDisplayAmount(
-              String(
-                fromSmallestUnit(awaitingTx.disbursedAmount || "0", awaitingTx.tokenDecimals ?? 6)
-              )
-            )}{" "}
-            {awaitingTx.token}
-            {awaitingTx.createdAt && (
-              <span className="text-amber-500 dark:text-amber-400/70">
-                {" "}
-                since {formatDate(awaitingTx.createdAt, "UTC")}
-              </span>
-            )}
-          </span>
-        </div>
-      )}
-
-      {/* Milestone completion summary */}
-      {milestoneSummary && (
-        <div className="flex items-center gap-4 text-xs text-gray-500 dark:text-zinc-400">
-          {invoiceRequired && (
-            <span>
-              <span className="font-medium text-gray-700 dark:text-zinc-300">
-                {milestoneSummary.received}/{milestoneSummary.total}
-              </span>{" "}
-              {milestoneSummary.total === 1 ? "invoice" : "invoices"} received
-            </span>
-          )}
-          <span>
-            <span className="font-medium text-gray-700 dark:text-zinc-300">
-              {milestoneSummary.paid}/{milestoneSummary.total}
-            </span>{" "}
-            {milestoneSummary.total === 1 ? "milestone" : "milestones"} paid
-          </span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── Milestones Section ───────────────────────────────────────────────────────
-
-interface MilestonesSectionProps {
-  grant: ProjectDetailsSidebarGrant;
-  milestoneInvoices: CommunityPayoutInvoiceInfo[];
-  invoiceRequired: boolean;
-  milestoneEdits: Record<
-    string,
-    { invoiceReceivedAt?: string | null; milestoneUID?: string | null }
-  >;
-  allocationByUID: Map<string, string>;
-  todayLocal: string;
-  getMilestoneKey: (invoice: CommunityPayoutInvoiceInfo, idx: number) => string;
-  getInvoiceReceivedDate: (invoice: CommunityPayoutInvoiceInfo, idx: number) => string | null;
-  handleInvoiceReceivedDateChange: (
-    milestoneKey: string,
-    milestoneUID: string | null,
-    dateValue: string
-  ) => void;
-}
-
-function MilestonesSection({
-  grant,
-  milestoneInvoices,
-  invoiceRequired,
-  milestoneEdits,
-  allocationByUID,
-  todayLocal,
-  getMilestoneKey,
-  getInvoiceReceivedDate,
-  handleInvoiceReceivedDateChange,
-}: MilestonesSectionProps) {
-  return (
-    <div>
-      <h3 className="text-sm font-semibold text-gray-900 dark:text-zinc-100 mb-1">
-        Milestones ({milestoneInvoices.length})
-      </h3>
-      {invoiceRequired && (
-        <p className="text-xs text-gray-400 dark:text-zinc-500 mb-3">
-          Set the received date to update invoice status. Save changes when done.
-        </p>
-      )}
-
-      {milestoneInvoices.length === 0 ? (
-        <p className="text-sm text-gray-500 dark:text-zinc-400 py-4 text-center">
-          No milestones configured yet.
-        </p>
-      ) : (
-        <TooltipProvider delayDuration={150}>
-          <div className="overflow-x-auto rounded-lg border border-gray-200 dark:border-zinc-700">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="bg-gray-50 dark:bg-zinc-800 border-b border-gray-200 dark:border-zinc-700">
-                  <th className="text-left py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[140px]">
-                    Milestone
-                  </th>
-                  <th className="text-center py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[100px]">
-                    Milestone Status
-                  </th>
-                  <th className="text-right py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[80px]">
-                    Allocation
-                  </th>
-                  {invoiceRequired && (
-                    <th className="text-center py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[110px]">
-                      Invoice Status
-                    </th>
-                  )}
-                  {invoiceRequired && (
-                    <th className="text-center py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[130px]">
-                      Received
-                    </th>
-                  )}
-                  <th className="text-center py-3 px-3 font-medium text-gray-600 dark:text-zinc-400 min-w-[110px]">
-                    Payment
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="divide-y divide-gray-100 dark:divide-zinc-800">
-                {milestoneInvoices.map((invoice, idx) => {
-                  const mKey = getMilestoneKey(invoice, idx);
-                  const receivedDateValue = getInvoiceReceivedDate(invoice, idx);
-                  const paymentCfg = paymentStatusConfig[invoice.paymentStatus ?? "unpaid"];
-                  const effectiveMsStatus = getEffectiveMilestoneStatus(
-                    invoice.milestoneStatus,
-                    invoice.milestoneDueDate
-                  );
-                  const msCfg =
-                    milestoneStatusConfig[effectiveMsStatus] ??
-                    milestoneStatusConfig[MilestoneLifecycleStatus.PENDING];
-                  const msTooltip = getMilestoneStatusTooltip(
-                    effectiveMsStatus,
-                    invoice.milestoneStatusUpdatedAt,
-                    invoice.milestoneDueDate
-                  );
-                  const invStatusKey = getInvoiceStatusKey(invoice.invoiceStatus);
-                  const invCfg = invoiceStatusConfig[invStatusKey];
-                  const isEdited = milestoneEdits[mKey] !== undefined;
-                  const isCleared =
-                    isEdited &&
-                    milestoneEdits[mKey]?.invoiceReceivedAt === null &&
-                    invoice.invoiceReceivedAt !== null;
-
-                  return (
-                    <tr
-                      key={invoice.milestoneUID || `${invoice.milestoneLabel}-${idx}`}
-                      className={cn(
-                        "transition-colors",
-                        isCleared
-                          ? "bg-amber-50/50 dark:bg-amber-900/10"
-                          : isEdited
-                            ? "bg-blue-50/50 dark:bg-blue-900/10"
-                            : "bg-white dark:bg-zinc-950"
-                      )}
-                    >
-                      <td className="py-3 px-3">
-                        <div className="flex items-center gap-1.5">
-                          {isEdited && (
-                            <span
-                              className={cn(
-                                "h-1.5 w-1.5 rounded-full shrink-0",
-                                isCleared ? "bg-amber-400" : "bg-blue-400"
-                              )}
-                            />
-                          )}
-                          <span className="font-medium text-gray-900 dark:text-zinc-100 line-clamp-2">
-                            {invoice.milestoneLabel}
-                          </span>
-                        </div>
-                      </td>
-                      <td className="py-3 px-3 text-center">
-                        <Tooltip>
-                          <TooltipTrigger asChild>
-                            <span
-                              className={cn(
-                                "inline-flex items-center gap-1.5 text-xs font-medium cursor-default",
-                                msCfg.textColor
-                              )}
-                            >
-                              <span
-                                className={cn("h-1.5 w-1.5 rounded-full shrink-0", msCfg.dotColor)}
-                              />
-                              {msCfg.label}
-                            </span>
-                          </TooltipTrigger>
-                          <TooltipContent>
-                            <p>{msTooltip}</p>
-                          </TooltipContent>
-                        </Tooltip>
-                        {effectiveMsStatus === MilestoneLifecycleStatus.COMPLETED && (
-                          <Tooltip>
-                            <TooltipTrigger asChild>
-                              <ExclamationTriangleIcon className="inline h-3.5 w-3.5 ml-1 text-amber-500 cursor-default" />
-                            </TooltipTrigger>
-                            <TooltipContent>
-                              <p>The milestone has not been verified yet</p>
-                            </TooltipContent>
-                          </Tooltip>
-                        )}
-                      </td>
-                      <td className="py-3 px-3 text-right">
-                        {(() => {
-                          const amount =
-                            invoice.allocatedAmount ??
-                            (invoice.milestoneUID
-                              ? allocationByUID.get(invoice.milestoneUID)
-                              : undefined) ??
-                            null;
-                          if (amount !== null) {
-                            return (
-                              <span className="text-xs tabular-nums text-gray-700 dark:text-zinc-300">
-                                {formatDisplayAmount(amount)}
-                                {grant.currency && (
-                                  <span className="text-gray-400 dark:text-zinc-500 ml-0.5">
-                                    {grant.currency}
-                                  </span>
-                                )}
-                              </span>
-                            );
-                          }
-                          return (
-                            <span className="text-xs text-gray-300 dark:text-zinc-600">
-                              &mdash;
-                            </span>
-                          );
-                        })()}
-                      </td>
-                      {invoiceRequired && (
-                        <td className="py-3 px-3 text-center">
-                          <span
-                            className={cn(
-                              "inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium",
-                              invCfg.className
-                            )}
-                          >
-                            {invCfg.label}
-                          </span>
-                        </td>
-                      )}
-                      {invoiceRequired && (
-                        <td className="py-3 px-3 text-center">
-                          <div className="flex flex-col items-center gap-0.5">
-                            <Input
-                              type="date"
-                              max={todayLocal}
-                              value={receivedDateValue ?? ""}
-                              onChange={(e) =>
-                                handleInvoiceReceivedDateChange(
-                                  mKey,
-                                  invoice.milestoneUID,
-                                  e.target.value
-                                )
-                              }
-                              className="h-7 text-xs w-[140px] mx-auto bg-white dark:bg-zinc-900"
-                              aria-label={`Invoice received date for ${invoice.milestoneLabel}`}
-                            />
-                            {invoice.invoiceReceivedBy && (
-                              <span
-                                className="text-[10px] text-gray-400 dark:text-zinc-500 font-mono"
-                                title={invoice.invoiceReceivedBy}
-                              >
-                                by {formatAddressForDisplay(invoice.invoiceReceivedBy)}
-                              </span>
-                            )}
-                          </div>
-                        </td>
-                      )}
-                      <td className="py-3 px-3 text-center">
-                        <div className="flex flex-col items-center gap-0.5">
-                          <span
-                            className={cn(
-                              "inline-flex items-center gap-1.5 text-xs font-medium",
-                              paymentCfg.textColor
-                            )}
-                          >
-                            <span
-                              className={cn(
-                                "h-1.5 w-1.5 rounded-full shrink-0",
-                                paymentCfg.dotColor
-                              )}
-                            />
-                            {paymentCfg.label}
-                          </span>
-                          {invoice.paymentStatusDate && (
-                            <span className="text-[10px] text-gray-400 dark:text-zinc-500 tabular-nums">
-                              {formatDate(invoice.paymentStatusDate, "UTC")}
-                            </span>
-                          )}
-                        </div>
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
-        </TooltipProvider>
-      )}
-    </div>
   );
 }
